Gestion de espacios y reservaciones con envio de correo

+++ database/migrations/2025_03_25_194319_reservaciones.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->time('hora_inicio');
             $table->time('hora_fin');
             $table->enum('estado', ['pendiente', 'confirmada', 'cancelada']);
-            $table->double('total_precio');
+            $table->decimal('total_precio',12,2);
             $table->timestamps();
         });
     }
